enzyme_neohookean test and new hpp file
